+++ doc/samples/qdiipred.py
@@ -5,7 +5,7 @@
 import pandas as pd
 import xalpha as xa
 
-xa.set_backend(backend="csv", path="../data", precached="20170103")
+xa.set_backend(backend="csv", path="../../../lof/data", precached="20200103")
 
 
 @xa.universal.lru_cache_time(ttl=60)
@@ -13,7 +13,7 @@ def cached_get_rt(code, **kws):
     return xa.get_rt(code, handler=False)
 
 
-@xa.universal.lru_cache_time(ttl=360)
+@xa.universal.lru_cache_time(ttl=1800)
 def cached_get_bar(code, *args, **kws):
     if code.startswith("commodities/"):
         return xa.get_bar(code, handler=False, *args, **kws)
@@ -28,9 +28,14 @@ qdiis = ["SH501018", "SZ160416", "SZ161129", "SZ160723", "SZ160216"]
 data = {"code": [], "t1": [], "t0": [], "t1rate": [], "t0rate": []}
 for c in qdiis:
     p = xa.QDIIPredict(c, fetch=True, save=True)
-    data["code"].append(c)
-    data["t1"].append(p.get_t1(return_date=False))
-    data["t1rate"].append(p.get_t1_rate(return_date=False))
-    data["t0"].append(p.get_t0(return_date=False))
-    data["t0rate"].append(p.get_t0_rate(return_date=False))
+    try:
+        data["t1"].append(p.get_t1(return_date=False))
+        data["t1rate"].append(p.get_t1_rate(return_date=False))
+        data["t0"].append(p.get_t0(return_date=False))
+        data["t0rate"].append(p.get_t0_rate(return_date=False))
+        data["code"].append(c)
+    except xa.exceptions.NonAccurate:
+        print("%s cannot be predicted exactly now" % c)
 df = pd.DataFrame(data)
+with open("demo.html", "w") as f:
+    df.to_html(f)