+++ xalpha/__init__.py
@@ -5,6 +5,7 @@ __name__ = "xalpha"
 import xalpha.policy
 import xalpha.remain
 import xalpha.misc
+import xalpha.exceptions
 from xalpha.evaluate import evaluate
 from xalpha.info import fundinfo, indexinfo, cashinfo, mfundinfo, FundReport
 from xalpha.multiple import mul, mulfix, imul
@@ -13,4 +14,11 @@ from xalpha.record import record, irecord
 from xalpha.trade import trade, itrade
 from xalpha.universal import get_daily, get_rt, get_bar, set_backend, set_handler, vinfo
 from xalpha.provider import show_providers, set_proxy
-from xalpha.toolbox import PEBHistory, Compare, SWPEBHistory, QDIIPredict, set_holdings
+from xalpha.toolbox import (
+    PEBHistory,
+    Compare,
+    SWPEBHistory,
+    QDIIPredict,
+    RTPredict,
+    set_holdings,
+)